Split web server up in classes
