+++ src/main/kotlin/ml/duncte123/skybot/objects/api/WebObjects.kt
@@ -44,7 +44,7 @@ open class ApiObject {
         this::class.memberProperties.forEach {
             if (it.visibility == KVisibility.PUBLIC) {
 
-                if (it.name == "file") {
+                if (it.name == "file" || it.name == "image") {
                     val file = it.getter.call(this) as String
                     json.put(it.name, "https://cdn.duncte123.me/" + file.eloquent())
                 } else {
